Sets db check on load_segments to datestop instead of datestart.
It is not uncommon to have a 4 day load segment at the end of the week
and to not approve loads until 3 days before the end of the week.
Checking that datestart of the last load segment is at least 1 day in
the future is prone to false positives.  Checking datestop should be
more reliable and informative for this data type.