Use age_str in hourly_watch
Trying to use the age attribute directly in the template
was problematic in the case when the file did not exists.
This new code does the "if exists" checking back in the Python.  The
age_str logic in jobwatch was not sufficient because these
needed unit conversion in the hourly_watch version.